+++ loader/include/Geode/utils/file.hpp
@@ -32,10 +32,10 @@ namespace geode::utils::file {
     template <class T>
     Result<T> readFromJson(std::filesystem::path const& file) {
         GEODE_UNWRAP_INTO(auto json, readJson(file));
-        if (!json.template is<T>()) {
+        if (!json.is<T>()) {
             return Err("JSON is not of type {}", typeid(T).name());
         }
-        return Ok(json.template as<T>());
+        return Ok(json.as<T>());
     }
 
     GEODE_DLL Result<> writeString(std::filesystem::path const& path, std::string const& data);